requirements.txt
langchain
langchain-openai
langchain-community
langchain-chroma
chromadb
openai
python-dotenv
